Working Gyro

+++ Main/Camera.h
@@ -4,10 +4,10 @@ class Camera
 {
   public:
   Camera(){};
-  Camera(uint8_t, uint8_t, uint8_t, uint8_t);
+  Camera(int, int, int, int);
   int GetColor();
   int GetLetter();
   private:
-  uint8_t color1, color2, letter1, letter2; 
+  int color1, color2, letter1, letter2; 
 };
 #endif
